Proof of concept for injecting new dungeon layouts
Just a very quick PoC to look at the possibility of changing dungeon layouts at runtime.
Parking these changes in a branch so they can be referred back to later. Do not merge back to master.
Ideally this feature would open up dungeon layouts to modding system in a similar way that terrain sampler opens heightmaps. Then it would be possible to create mods for smaller/larger dungeons with custom constraints.
Taking it completely next-level, it might even be possible to use custom blocks with all new layouts and assets for truly unique dungeons. That's obviously a long-term goal however.