remove ambiant light smoothing
Smoothing gives unexpected results; Say you exit a bright lighted room
to enter the streets at night, you'd expect to only see darkness at
first, then start to recognize the street as your eyes accomodate.
It's purpose was probably not to emulate eyes accomodation, but the
problem is that it gives the opposite results.
Let's delegate eyes accomodation simulation to post-processing (if one
wishes to use that), and simply remove smoothing.