cursor mode prevents camera rotation
Cursor mode prevented mouse control from rotating the camera, but they
were other causes for camera rotation: head bobbing and camera recoil.

I cancelled camera rotation from head bobbing;
Recoil is a little different, because it's purely rotation, so a goal
would be that when recoils end the camera direction is unchanged; Little
is necessary for that, because it's already made of full half periods of
a dampened sinusoid: first, full periods must be used, and second
dampening should only happen in steps with each period so that the
movement is perfectly symetric.

It should be noted that removing rotation from head bobbing is hardly
noticeable; And that recoil is a *lot* more effective in cursor mode.
I suspect mouse look filters most of camera rotations.

Forums: https://forums.dfworkshop.net/viewtopic.php?f=5&t=2776