Improved workflow for retro postprocess
@petchema This seems like the cleanest approach by far:

1. Render world and sky to separate render targets. Both targets use camera clearing to colour and depth.
2. At final composite time, first draw sky, then draw world over sky.
3. Using a simple depth-clip colour-copy shader to filter world fragments untouched by frame process. Sky remains visible and untouched by postprocess behind world render.

Because this is clipping by depth == 0 rather than some alpha threshold, we don't need to do anything special with render queues to fix self-shadowing. Still need to solve weather particles though. Haven't looked into that yet.

Pushing this to a new branch for your consideration. As with the last approach, the equivalent depth check would need to be built into postprocess shader for blit. But once we have it working, we shouldn't need to hack around much if at all with other shaders and systems. It's all nice and self-contained to retro rendering.