Inversion of control using functionnal approach
Don't send hero gender to the grammar processor in each place it
changes, but instead give the grammar processor the getters to access
that information from the game.

(setting the function in StartGameBehavior.cs sucks, I don't know where
to put code that would be run no matter how the game is started).