Randomize enemies orientation
It seems classic spawns all enemies facing north, and that's also what
Daggerfall Unity does.

However, are the senses of standing enemies equally isotropic in classic
and DFU? If not, then this could actually make the games behave
differently.

In any case, try placing enemies with a random orientation just for
experimentation.